Fix more comments WIP
