+++ app/web/features/dashboard/UpcomingStays.tsx
@@ -1,19 +1,13 @@
 import {
   ArrowBack,
   ArrowForward,
-  EventOutlined,
   Luggage,
   MeetingRoom,
-  PlaceOutlined,
 } from "@mui/icons-material";
-import { Box, IconButton, Skeleton, styled, Typography } from "@mui/material";
-import { alpha } from "@mui/material/styles";
+import { Box, IconButton, styled, Typography } from "@mui/material";
 import { useQuery } from "@tanstack/react-query";
 import Alert from "components/Alert";
-import Avatar from "components/Avatar";
-import { useAuthContext } from "features/auth/AuthProvider";
 import { hostRequestsListKey } from "features/queryKeys";
-import { useLiteUser } from "features/userQueries/useLiteUsers";
 import { RpcError } from "grpc-web";
 import { useTranslation } from "i18n";
 import { DASHBOARD } from "i18n/namespaces";
@@ -21,15 +15,12 @@ import Link from "next/link";
 import { HostRequestStatus } from "proto/conversations_pb";
 import { HostRequest, ListHostRequestsRes } from "proto/requests_pb";
 import { useEffect, useRef, useState } from "react";
-import { routeToEditProfile, routeToHostRequest, searchRoute } from "routes";
+import { routeToEditProfile, searchRoute } from "routes";
 import { service } from "service";
 import { theme } from "theme";
-import {
-  localizeDateTimeRange,
-  localizeRelativeDays,
-  UTC_TIMEZONE,
-} from "utils/date";
 import dayjs from "utils/dayjs";
+
+import UpcomingStayCard, { UpcomingStayCardSkeleton } from "./UpcomingStayCard";
 
 const CARD_WIDTH = 220;
 const CARD_GAP = 12;
@@ -42,14 +33,9 @@ const ACCEPTED_STATUSES = new Set([
 function filterUpcoming(
   requests: HostRequest.AsObject[],
 ): HostRequest.AsObject[] {
-  const today = dayjs().startOf("day");
   return requests
-    .filter(
-      (r) =>
-        ACCEPTED_STATUSES.has(r.status) &&
-        dayjs(r.fromDate).diff(today, "day") >= 0,
-    )
-    .sort((a, b) => dayjs(a.fromDate).diff(dayjs(b.fromDate)));
+    .filter((r) => ACCEPTED_STATUSES.has(r.status))
+    .sort((a, b) => dayjs.utc(a.fromDate).diff(dayjs.utc(b.fromDate)));
 }
 
 const SectionHeader = styled("div")({
@@ -69,74 +55,6 @@ const Track = styled(Box)({
   "&::-webkit-scrollbar": { display: "none" },
 });
 
-const StyledCard = styled(Box)(({ theme }) => ({
-  display: "flex",
-  flexDirection: "column",
-  padding: theme.spacing(1.5),
-  border: "1px solid var(--mui-palette-divider)",
-  borderRadius: 10,
-  background: "var(--mui-palette-background-paper)",
-  cursor: "pointer",
-  transition: "border-color 0.2s, box-shadow 0.2s",
-  height: "100%",
-  minWidth: 0,
-  overflow: "hidden",
-  boxSizing: "border-box",
-  "&:hover": {
-    borderColor: "var(--mui-palette-primary-main)",
-    boxShadow: "0 2px 8px rgba(0,0,0,0.08)",
-  },
-}));
-
-const DaysChip = styled("span")<{ imminent: boolean }>(
-  ({ theme, imminent }) => ({
-    fontSize: 11,
-    fontWeight: 700,
-    borderRadius: 999,
-    padding: "2px 8px",
-    whiteSpace: "nowrap",
-    flexShrink: 0,
-    marginLeft: "auto",
-    background: imminent
-      ? alpha(theme.palette.secondary.main, 0.12)
-      : "var(--mui-palette-grey-50)",
-    color: imminent
-      ? "var(--mui-palette-secondary-dark)"
-      : "var(--mui-palette-text-secondary)",
-  }),
-);
-
-const IdentityRow = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  gap: theme.spacing(1.25),
-}));
-
-const TextBlock = styled("div")({
-  minWidth: 0,
-});
-
-const MetaRow = styled("div")(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "space-between",
-  gap: theme.spacing(1),
-  marginTop: theme.spacing(1.5),
-  paddingTop: theme.spacing(1.25),
-  borderTop: "1px dashed var(--mui-palette-divider)",
-}));
-
-const MetaDate = styled("div")(({ theme }) => ({
-  display: "inline-flex",
-  alignItems: "center",
-  gap: theme.spacing(0.75),
-  fontSize: 12,
-  fontWeight: 500,
-  color: "var(--mui-palette-text-primary)",
-  minWidth: 0,
-  overflow: "hidden",
-}));
-
 const EmptyStateRow = styled("div")(({ theme }) => ({
   display: "flex",
   alignItems: "center",
@@ -146,170 +64,6 @@ const EmptyStateRow = styled("div")(({ theme }) => ({
   borderRadius: 10,
   background: "var(--mui-palette-grey-50)",
 }));
-
-// ── UpcomingStayCard ───────────────────────────────────────────────────────
-
-interface UpcomingStayCardProps {
-  hostRequest: HostRequest.AsObject;
-}
-
-function daysUntilLabel(days: number, locale: string): string {
-  return localizeRelativeDays(days, locale);
-}
-
-function UpcomingStayCard({ hostRequest }: UpcomingStayCardProps) {
-  const { authState } = useAuthContext();
-  const {
-    t,
-    i18n: { language: locale },
-  } = useTranslation([DASHBOARD]);
-
-  const isHost = authState.userId === hostRequest.hostUserId;
-  const otherUserId = isHost
-    ? hostRequest.surferUserId
-    : hostRequest.hostUserId;
-  const { data: otherUser, isLoading } = useLiteUser(otherUserId);
-
-  const fromDate = dayjs.tz(hostRequest.fromDate, UTC_TIMEZONE);
-  const toDate = dayjs.tz(hostRequest.toDate, UTC_TIMEZONE);
-  const nights = toDate.diff(fromDate, "day");
-  const daysUntil = fromDate.diff(dayjs().startOf("day"), "day");
-  const isImminent = daysUntil <= 3;
-
-  const dateRange = localizeDateTimeRange(fromDate, toDate, {
-    timezone: UTC_TIMEZONE,
-    locale,
-    includeTime: false,
-    abbreviate: true,
-  });
-
-  const primary = isLoading ? (
-    <Skeleton width={80} />
-  ) : (
-    (otherUser?.name ?? "—")
-  );
-
-  return (
-    <Link
-      href={routeToHostRequest(hostRequest.hostRequestId)}
-      style={{
-        display: "flex",
-        textDecoration: "none",
-        color: "inherit",
-        height: "100%",
-        width: "100%",
-      }}
-    >
-      <StyledCard sx={{ flex: 1 }}>
-        <IdentityRow>
-          <Avatar user={otherUser} isProfileLink={false} />
-          <TextBlock>
-            <Typography
-              variant="h3"
-              noWrap
-              sx={{ fontSize: 14, fontWeight: 700, lineHeight: 1.3 }}
-            >
-              {primary}
-            </Typography>
-            <Typography
-              variant="body2"
-              sx={{
-                fontSize: 12,
-                color: "var(--mui-palette-text-secondary)",
-                display: "flex",
-                alignItems: "center",
-                gap: "3px",
-                overflow: "hidden",
-              }}
-            >
-              {isLoading ? (
-                <Skeleton width={60} />
-              ) : (
-                <>
-                  <PlaceOutlined style={{ fontSize: 12, flexShrink: 0 }} />
-                  <span
-                    style={{
-                      overflow: "hidden",
-                      textOverflow: "ellipsis",
-                      whiteSpace: "nowrap",
-                    }}
-                  >
-                    {otherUser?.city}
-                  </span>
-                </>
-              )}
-            </Typography>
-          </TextBlock>
-          {isImminent && (
-            <DaysChip imminent>{daysUntilLabel(daysUntil, locale)}</DaysChip>
-          )}
-        </IdentityRow>
-        <MetaRow>
-          <MetaDate>
-            <EventOutlined sx={{ fontSize: 13, flexShrink: 0 }} />
-            <span
-              style={{
-                overflow: "hidden",
-                textOverflow: "ellipsis",
-                whiteSpace: "nowrap",
-              }}
-            >
-              {dateRange}
-            </span>
-          </MetaDate>
-          <Typography
-            component="span"
-            variant="body2"
-            sx={{
-              fontSize: 12,
-              color: "var(--mui-palette-text-secondary)",
-              flexShrink: 0,
-            }}
-          >
-            {t("dashboard:stays.night_count", { count: nights })}
-          </Typography>
-        </MetaRow>
-      </StyledCard>
-    </Link>
-  );
-}
-
-function UpcomingStayCardSkeleton() {
-  return (
-    <Box
-      sx={{
-        p: 1.5,
-        border: "1px solid var(--mui-palette-divider)",
-        borderRadius: "10px",
-        background: "var(--mui-palette-background-paper)",
-      }}
-    >
-      <Box sx={{ display: "flex", gap: 1.25, pr: 7.5, mb: 1.5 }}>
-        <Skeleton
-          variant="circular"
-          width={40}
-          height={40}
-          sx={{ flexShrink: 0 }}
-        />
-        <Box sx={{ flex: 1 }}>
-          <Skeleton height={16} sx={{ mb: 0.5 }} />
-          <Skeleton height={12} width="60%" />
-        </Box>
-      </Box>
-      <Box
-        sx={{
-          display: "flex",
-          justifyContent: "space-between",
-          pt: 1.25,
-          borderTop: "1px dashed var(--mui-palette-divider)",
-        }}
-      >
-        <Skeleton height={14} width={90} />
-        <Skeleton height={14} width={50} />
-      </Box>
-    </Box>
-  );
-}
 
 // ── UpcomingStaysWidget ────────────────────────────────────────────────────
 
@@ -332,6 +86,7 @@ function UpcomingStaysWidget({
   emptyCtaLabel,
   emptyCtaHref,
 }: UpcomingStaysWidgetProps) {
+  const { t } = useTranslation([DASHBOARD]);
   const scrollerRef = useRef<HTMLDivElement>(null);
   const [canScrollLeft, setCanScrollLeft] = useState(false);
   const [canScrollRight, setCanScrollRight] = useState(false);
@@ -339,8 +94,12 @@ function UpcomingStaysWidget({
   const updateScrollState = () => {
     const el = scrollerRef.current;
     if (!el) return;
-    setCanScrollLeft(el.scrollLeft > 4);
-    setCanScrollRight(el.scrollLeft < el.scrollWidth - el.clientWidth - 4);
+    // scrollLeft: pixels scrolled; clientWidth: visible width; scrollWidth: total content width.
+    // scrollLeft is a float on high-DPI screens, so Math.round prevents the right arrow staying lit at the end.
+    setCanScrollLeft(el.scrollLeft > 0);
+    setCanScrollRight(
+      Math.round(el.scrollLeft) < el.scrollWidth - el.clientWidth,
+    );
   };
 
   useEffect(() => {
@@ -387,7 +146,7 @@ function UpcomingStaysWidget({
             onClick={() => scroll(-1)}
             disabled={!canScrollLeft}
             color={canScrollLeft ? "primary" : "default"}
-            aria-label="Previous"
+            aria-label={t("dashboard:prev_page_button_a11y")}
           >
             <ArrowBack fontSize="small" />
           </IconButton>
@@ -396,7 +155,7 @@ function UpcomingStaysWidget({
             onClick={() => scroll(1)}
             disabled={!canScrollRight}
             color={canScrollRight ? "primary" : "default"}
-            aria-label="Next"
+            aria-label={t("dashboard:next_page_button_a11y")}
           >
             <ArrowForward fontSize="small" />
           </IconButton>
@@ -430,7 +189,7 @@ function UpcomingStaysWidget({
               component="span"
               variant="body2"
               sx={{
-                color: "var(--mui-palette-primary-dark)",
+                color: "var(--mui-palette-primary-main)",
                 fontWeight: 600,
                 "&:hover": { textDecoration: "underline" },
               }}
@@ -474,6 +233,7 @@ export default function UpcomingStays() {
       service.requests.listHostRequests({
         type: "surfing",
         count: 20,
+        onlyActive: true,
       }),
   });
 
@@ -487,6 +247,7 @@ export default function UpcomingStays() {
       service.requests.listHostRequests({
         type: "hosting",
         count: 20,
+        onlyActive: true,
       }),
   });
