+++ app/web/features/dashboard/locales/en.json
@@ -56,13 +56,13 @@
   "show_map": "Show Map",
   "hero_image_alt": "Hot-air balloons in Capadocia",
   "hero_image_attribution": "Photo by <1>Mesut Kaya</1> on <3>Unsplash</3>",
+  "prev_page_button_a11y": "Previous page",
+  "next_page_button_a11y": "Next page",
   "discussions": {
     "community_header": "Discussions in your communities",
     "community_empty_message": "No discussions yet in your communities.",
     "replies_count_label_one": "{{count}} reply",
-    "replies_count_label_other": "{{count}} replies",
-    "prev_page_label": "Previous page",
-    "next_page_label": "Next page"
+    "replies_count_label_other": "{{count}} replies"
   },
   "events": {
     "your_upcoming_header": "Your upcoming events",
